requirements.txt
beautifulsoup4==4.6.3
Django==2.1.4
django-background-tasks==1.2.0
django-compat==1.0.15
django-cors-headers==2.4.0
djangorestframework==3.9.0
mysqlclient==1.3.14
pytz==2018.7
six==1.12.0
